% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_beeping.R
\name{beeping}
\alias{beeping}
\title{beeping tag to beep when instruction finishes}
\usage{
beeping(.sound = 1, .expr = NULL)
}
\arguments{
\item{.sound}{character string or number specifying what sound to be played,
forwarded to the \code{sound} argument of \code{beepr::beep}}

\item{.expr}{an optional expression to be excecuted before the sound,
forwarded to the \code{expr} argument of \code{beepr::beep}}
}
\description{
uses \code{beepr::beep} to signal that the call was executed.
}
\examples{
\dontrun{
beeping$paste("hello","world")
beeping(2)$paste("hello","world")
beeping("facebook")$paste("hello","world")
}
}
