% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tidy.R
\name{bng}
\alias{bng}
\title{bng tag to enable tidy evaluation or arguments}
\usage{
bng()
}
\description{
Use tidy evaluation on all arguments
}
\examples{
u <- "speed"
v <- quote(dist)
w <- rlang::quo(time)
x <- list(a=c(1, 2), b=c(3, 4))
bng$transform(head(cars,2), !!w := !!v / !!rlang::sym(u), !!!x)
}
