% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_checking_args.R
\name{checking_args}
\alias{checking_args}
\title{checking_args to test arguments before running a call}
\usage{
checking_args(.check, .message = NULL)
}
\arguments{
\item{.check}{an expression that should be \code{TRUE} for the call to proceed}

\item{.message}{An error message, falls back on a default message if NULL}
}
\description{
checking_args can make checks more readable and avoid defining temporary
variables
}
\examples{
\dontrun{
mean(1:5, na.rm = 1)
checking_args(is.logical(na.rm))$mean(1:5, na.rm = 1)
checking_args(is.logical(na.rm),"na.rm should be logical!")$mean(1:5, na.rm = 1)
}
}
