% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_composing.R
\name{composing}
\alias{composing}
\title{composing tag to start a dollar piped functional sequence}
\arguments{
\item{x}{a function}
}
\value{
a tagged function
}
\description{
\code{composing} is a tag wrapped around \emph{magrittr}'s functional sequence
feature. It works in similar ways except that it uses \code{$} rather than
\code{\%>\%} and doesn't start with a dot.
}
\examples{
\dontrun{
composing()(head)$dim()(iris)
composing$head(2)$gsub("h","X",.)(c("hello","hi","foo"))
}
}
