% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_error_not_found.R
\name{conditionMessage.TAGS_error_not_found}
\alias{conditionMessage.TAGS_error_not_found}
\title{@title file_error_not_found

@export}
\usage{
\method{conditionMessage}{TAGS_error_not_found}(c)
}
\description{
@title file_error_not_found

@export
}
