% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_files.R
\name{create_fake_yml_object}
\alias{create_fake_yml_object}
\title{@title create_fake_yaml_object
@description   GOAL:   create 4 use cases to test TAGS: ....
(1) all correct
(2) no yaml of any kind, no TAGS
(3) has yaml, no TAGS: line
(4) has yaml, has TAGS:  but no acutal tags (blank)
@examples
\dontrun{
    h  <- create_fake_yml_object(the_tags_line = "tag1,tag2, tag3")
    class(h)
    str(h)
    ls.str(h)
    }}
\usage{
create_fake_yml_object(the_tags_line = NULL)
}
\description{
@title create_fake_yaml_object
@description   GOAL:   create 4 use cases to test TAGS: ....
(1) all correct
(2) no yaml of any kind, no TAGS
(3) has yaml, no TAGS: line
(4) has yaml, has TAGS:  but no acutal tags (blank)
@examples
\dontrun{
    h  <- create_fake_yml_object(the_tags_line = "tag1,tag2, tag3")
    class(h)
    str(h)
    ls.str(h)
    }
}
