% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{fetch_yaml}
\alias{fetch_yaml}
\title{fetch_yaml}
\usage{
fetch_yaml(x)
}
\arguments{
\item{x}{chr\link{1}}
}
\value{
char[] everything between --- inclusive, char vector
}
\description{
fetch_yaml
}
\keyword{internal}
