% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{@title get_files}
\usage{
get_files(path = ".", pattern = NULL, recursive = FALSE)
}
\description{
@description  Given a path and pattern, return char vector of file names.
Ignores files that begin with '_'
@param path  Directory to query.
@param pattern regex pattern to match files
@return full file name (with directory)
@export
}
