% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_header}
\alias{get_header}
\title{@title get_header
@description   Wrapper for yaml_front_matter.  Removes "---"; returns list.
ie header$TAGS (if exists) and contents are char\link{1}}
\usage{
get_header(files = NULL)
}
\description{
@param files. char[] of file names.
@return char\link{1} For ONE file, returns yaml header.
}
