% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_t}
\alias{get_t}
\title{get_t}
\usage{
get_t(e)
}
\description{
returns TAGS
\subsection{the_yaml is named list.  Contents of each element is list of length n,}{
}

\subsection{correspondng to number of lines of YAML. Content of this element is chr()}{
}

\subsection{YAML.  If a file had no YAML, then it is not here.}{
}

\subsection{In this function, return the TAGS or NA.}{
}

\subsection{}{
}
}
