% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_TAGS}
\alias{get_TAGS}
\title{@title get_TAGS}
\usage{
get_TAGS(header)
}
\arguments{
\item{header}{For one file, header is named list. Each name entry represents
a single yaml line and its contents.}
}
\value{
char\link{1} that is content of TAGS:
}
\description{
@description Given one yaml header (everything between \verb{---} inclusive), return the line begins with \verb{TAGS:}  Return NA if no line contains \verb{TAGS:}.
}
