% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_yaml}
\alias{get_yaml}
\title{get_yaml}
\usage{
get_yaml(file = NULL, dir = NULL)
}
\arguments{
\item{file}{char\link{1} of file name.}

\item{dir}{character\link{1} full path (without filename)}
}
\value{
unnamed list.  List contents char[] of length equal to lines of
YAML.
}
\description{
Wrapper for yaml_front_matter.  Given a file containing yaml,
removes "---"; returns named list with one element.  The list contains
character[] corresponding to ONE line yaml.  If file contains no yaml, it
is skipped.
}
