% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tidy.R
\name{grouping_by}
\alias{grouping_by}
\title{grouping_by tag to group and ungroup around another operation}
\usage{
grouping_by(.by)
}
\arguments{
\item{.by}{variables to group by, either as a character vector or a call to
\code{vars}}
}
\description{
The tagged function gains a parameter \code{.by}, groups that existed before
are preserved. One can group by a computed variable, it will be kept if
named, dropped otherwise.
}
\examples{
\dontrun{
library(dplyr)
grouping_by$summarize(iris,meanSL = mean(Sepal.Length), .by="Species")
grouping_by(.by="Species")$summarize(iris,meanSL = mean(Sepal.Length))
grouping_by$summarize_all(iris, mean, .by="Species")
grouping_by$slice(iris,1, .by="Species")
# unnamed groupings are not kept
grouping_by$summarize(
  iris,meanSW = mean(Sepal.Width), .by= vars(Species, Sepal.Width > 3.2))
# named groupings are kept
grouping_by$summarize(
  iris,meanSW = mean(Sepal.Width), .by= vars(Species, long_sepal = Sepal.Width > 3.2))
  }
}
