% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tidy.R
\name{grp}
\alias{grp}
\title{grp tag to group and ungroup around another operation}
\usage{
grp(.by)
}
\arguments{
\item{.by}{variables to group by, either as a character vector or a call to
\code{vars}}
}
\description{
The tagged function gains a parameter \code{.by}, groups that existed before
are preserved. One can group by a computed variable, it will be kept if
named, dropped otherwise.
}
\examples{
library(dplyr)
grp$summarize(iris,meanSL = mean(Sepal.Length), .by="Species")
grp(.by="Species")$summarize(iris,meanSL = mean(Sepal.Length))
grp$summarize_all(iris, mean, .by="Species")
grp$slice(iris,1, .by="Species")
# unnamed groupings are not kept
grp$summarize(
  iris,meanSW = mean(Sepal.Width), .by= vars(Species, Sepal.Width > 3.2))
# named groupings are kept
grp$summarize(
  iris,meanSW = mean(Sepal.Width), .by= vars(Species, long_sepal = Sepal.Width > 3.2))
}
