% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_using_a_new_session.R
\name{in_parralel}
\alias{in_parralel}
\title{in_parralel tag to execute a call in another session}
\usage{
in_parralel(.wait = TRUE, .path = getwd())
}
\arguments{
\item{.wait}{wether we wait for the result}

\item{.path}{path where all temp files are stored}
}
\description{
by executing a call in another session we gain the opporunity of displaying
dynamically the status, break the action after a timeout, or run the code
asynchronously and keep going.
}
\details{
Draft, ellipsis is not supported and might be easy to break.
}
\examples{
u <- cars
v <- 3
fun <- function(y,m){
  Sys.sleep(5)
  head(y, m)
}
x <- in_parralel(.wait=FALSE)$fun(u,v)
x
y <- in_parralel$fun(u,v)
y
}
