% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basics.R
\name{logging}
\alias{logging}
\title{logging tag to log call and time it took}
\usage{
logging(.time = TRUE, .print = TRUE)
}
\arguments{
\item{.time}{logical, wether to print the time}

\item{.print}{logical, wether to print the object}
}
\description{
logging tag to log call and time it took
}
\examples{
logging$head(cars,2)
}
