% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_dot_operations.R
\name{modifying_dots}
\alias{modifying_dots}
\alias{selecting_dots}
\alias{renaming_dots}
\alias{mutating_dots}
\alias{transmuting_dots}
\alias{reversing_dots}
\title{modifying_dots}
\usage{
selecting_dots(...vars)

renaming_dots(...vars)

mutating_dots(...exprs)

transmuting_dots(...exprs)

reversing_dots()
}
\arguments{
\item{...vars}{A list of columns generated by vars()}

\item{...exprs}{A list of expressions generated by exprs()}
}
\description{
tags inspired by dplyr functions to modify the dots of the call
}
\examples{
# the functions can be used in regular calls
reversing_dots$paste0("a", "b")

selecting_dots(vars(b2,starts_with("a")))$
  transform(head(cars,2), a1 = 1, b1 = 2, a2 = 3, b2 = speed *100)

mutating_dots(exprs(foo = 1000 * bar))$
  mutate(head(cars,2), bar = speed * 10)

# or with forwarded dots
fun <- function(...){
  print(
    selecting_dots(vars(b2,starts_with("a")))$
      mutate(head(cars,2), ...))
  print(
    renaming_dots(vars(foo = a1))$
      mutate(head(cars,2), ...))
  print(
    mutating_dots(exprs(a2 = 100*a2 + a1, foo = 1000 * b1))$
      mutate(head(cars,2), ...))
  print(
    transmuting_dots(exprs(a2 = 100*a2 + a1, foo = 1000 * b1))$
      mutate(head(cars,2), ...))
  print(
    reversing_dots$mutate(head(cars,2), ...))
}
fun(a1 = 1, b1 = 2, a2 = 3, b2 = 40)
}
