% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_base.R
\name{negating}
\alias{negating}
\title{negating tag to negate the input function}
\usage{
negating()
}
\description{
tag counterpart to \code{base::Negate}, and a wrapper around it.
}
\examples{
negating$is.character(1)
negating$is.numeric(1)
}
