% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tidy.R
\name{on_self}
\alias{on_self}
\title{on_self tag to allow self referencing in arguments}
\usage{
on_self()
}
\description{
Enhance functions like \code{transform}, \code{mutate}, \code{summarize} ... by allowing
arguments to reference themselves when they are formulas.
}
\details{
to detect rigorously the type of the argument it would be necessary to
evaluate it, which for these functions is not an option, therefore we
just check if the expressions starts with a \code{~} character.
}
\examples{
on_self$transform(head(iris,2), Petal.Width = ~1000*(.), Species = ~toupper(.))
on_self$summarize(iris, Petal.Width = ~median(.), Sepal.Length = ~mean(.))
}
