% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_beeping.R
\name{popping_up}
\alias{popping_up}
\title{popping_up tag to trigger message box when instruction finishes}
\usage{
popping_up(.message = NULL, .caption = "")
}
\arguments{
\item{.message}{character. The information field of the dialog box.
Forwarded to the \code{message} argument of \code{tcltk::tk_messageBox}}

\item{.caption}{the caption on the widget displayed.
Forwarded to the \code{caption} argument of \code{tcltk::tk_messageBox}}
}
\description{
wraps \code{tcltk::tk_messageBox} to signal that the call was executed.
}
\examples{
\dontrun{
popping_up$paste("hello","world")
popping_up("a message", "a caption")$paste("hello","world")
}
}
