% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_purrr_adverbs.R
\name{possibly2}
\alias{possibly2}
\title{possibly2  tag}
\usage{
possibly2(otherwise, quiet = TRUE)
}
\arguments{
\item{otherwise}{Default value to use when an error occurs.}

\item{quiet}{Hide errors (\code{TRUE}, the default), or display them
as they occur?}
}
\description{
similar to \code{purrr::possibly}
}
\examples{
possibly2$log("a", otherwise = NA_real_)
possibly2$log(10, otherwise = NA_real_)
}
