% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_base.R
\name{preserving_attr}
\alias{preserving_attr}
\title{preserving_attr tag to preserve attributes}
\usage{
preserving_attr(.arg = 1, .incl_row.names = FALSE,
  .incl_class = FALSE, .incl_names = FALSE, .incl_dim = FALSE)
}
\arguments{
\item{.arg}{index or name of argument to preserve}

\item{.incl_row.names}{wether to include \code{row.names} in attributes to preserve}

\item{.incl_class}{wether to include \code{class} in attributes to preserve}

\item{.incl_names}{wether to include \code{names} in attributes to preserve}

\item{.incl_dim}{wether to include \code{dim} in attributes to preserve}
}
\description{
preserving_attr tag to preserve attributes
}
\examples{
preserving_attr(.incl_class = TRUE)(purrr::map_dfr)(head(iris,2),identity)
preserving_attr(.incl_class = TRUE, .incl_row.names = TRUE)$lapply(head(iris,2),identity)
}
