% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_purrr_adverbs.R
\name{quietly2}
\alias{quietly2}
\title{quietly2  tag}
\usage{
quietly2(.f)
}
\arguments{
\item{.f}{A function, formula, or vector (not necessarily atomic).

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it is
converted to an extractor function. Character vectors index by
name and numeric vectors index by position; use a list to index
by position and name at different levels. If a component is not
present, the value of \code{.default} will be returned.}
}
\description{
similar to \code{purrr::quietly}
}
\examples{
quietly2$sqrt(-1)
quietly2$sqrt(4)
}
