% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{remove_comma}
\alias{remove_comma}
\title{remove_comma
remove_commas}
\usage{
remove_comma(s = character())
}
\arguments{
\item{s}{is single character vector or string char\link{1}, where s is of the
from "a,b,d..."  and commas are to be removed.}
}
\value{
char vector of form:  c("a", "b", "c", "d"
}
\description{
LEGACY (no longer used)
}
\keyword{internal}
