% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{remove_prefix}
\alias{remove_prefix}
\title{remove_prefix
remove_prefix}
\usage{
remove_prefix(s = character())
}
\arguments{
\item{s}{is single character vector or string char\link{1}, where s is of the
from "X: a,b,d..."  and X:  (TAGS:) is to be removed.}
}
\value{
char\link{1}  without X:
}
\description{
LEGACY (no longer used)
}
\keyword{internal}
