% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tidy.R
\name{rw}
\alias{rw}
\title{rw tag to apply a transformation rowwise}
\usage{
rw()
}
\description{
rw tag to apply a transformation rowwise
}
\examples{
rw$mutate(head(iris,3),X = mean(c(Sepal.Length, Sepal.Width)))
mutate(head(iris,3),X = mean(c(Sepal.Length, Sepal.Width)))
}
