% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_purrr_adverbs.R
\name{safely2}
\alias{safely2}
\title{safely2  tag}
\usage{
safely2(otherwise = NULL, quiet = TRUE)
}
\arguments{
\item{otherwise}{Default value to use when an error occurs.}

\item{quiet}{Hide errors (\code{TRUE}, the default), or display them
as they occur?}
}
\description{
similar to \code{purrr::safely}
}
\examples{
safely2$log(10, otherwise = NA_real_)
safely2$log("a", otherwise = NA_real_)
safely2(otherwise = NA_real_)$log(10)
safely2(otherwise = NA_real_)$log("a")
}
