% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basics.R
\name{strictly}
\alias{strictly}
\title{strictly tag to adjust strictness of a call}
\usage{
strictly(warn = 2)
}
\arguments{
\item{warn}{numeric, forwarded to \code{options()}}
}
\description{
strictly tag to adjust strictness of a call
}
\examples{
try(strictly$sqrt(-1))
strictly(0)$sqrt(-1)
strictly(-1)$sqrt(-1)
}
