% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_tracing_enclosing.R
\name{tracing}
\alias{tracing}
\title{tracing tag}
\usage{
tracing(tracer = NULL, exit = NULL, at = 1, edit = FALSE)
}
\description{
Emulates the behavior of \code{base::trace}.
}
\details{
It's not a wrapper around \code{base::trace}, instead it modifies the body of the
input function. If the input function is an \emph{S3 generic} the effect of
the tracing can be overwritten by the argument values, better in this case
to use the relevant method as a function input.
}
\examples{
tracing(quote(print("hello")), exit = quote(print("good bye")))$mean(1:5)
}
