% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basics.R
\name{trying}
\alias{trying}
\title{trying tag to try a call or if failure call alternate call}
\usage{
trying(.else, .silent = FALSE)
}
\description{
The expression stored in \code{else} will be executed if
}
\examples{
trying$paste("hello","world", .else = "hi")
trying$paste("hello", world, .else = "hi")
trying$paste("hello", world, .else = "hi", .silent = TRUE)
}
