% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tidy.R
\name{using_lambda}
\alias{using_lambda}
\title{using_lambda tag to use formula notation on FUN arguments}
\usage{
using_lambda()
}
\description{
inspired by \code{gsubfn::fn}, \code{..fn} is similar but use \code{rlang::as_function}
while \emph{gsubfn} uses its own API.
}
\examples{
using_lambda$ave(c(1,2,4,3,NA),c(1,1,2,2,2),FUN = ~mean(.,na.rm=TRUE))
}
