% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_tidy.R
\name{using_rowwise}
\alias{using_rowwise}
\title{using_rowwise tag to apply a transformation rowwise}
\usage{
using_rowwise()
}
\description{
using_rowwise tag to apply a transformation rowwise
}
\examples{
\dontrun{
library(dplyr)
using_rowwise$mutate(head(iris,3),X = mean(c(Sepal.Length, Sepal.Width)))
mutate(head(iris,3),X = mean(c(Sepal.Length, Sepal.Width)))
}
}
