% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_purrr_adverbs.R
\name{using_safely}
\alias{using_safely}
\alias{using_quietly}
\alias{using_possibly}
\title{3 functions
using_safely tag}
\usage{
using_safely(.otherwise = NULL, .quiet = TRUE)

using_quietly()

using_possibly(.otherwise, .quiet = TRUE)
}
\arguments{
\item{.otherwise}{Default value to use when an error occurs.}

\item{.quiet}{Hide errors (\code{TRUE}, the default), or display them as they occur?}
}
\description{
tag ounterparts of \emph{purrr} adverbs.
These are not wrappers around \emph{purrr} functions but contain some of their
code, and have a similar interface).
\code{trying} is an alias for \code{using_possibly}.

similar to \code{purrr::quietly}

similar to \code{purrr::possibly}
}
\details{
similar to \code{purrr::safely}
}
\examples{
using_safely$log(10, .otherwise = NA_real_)
using_safely$log("a", .otherwise = NA_real_)
using_safely(.otherwise = NA_real_)$log(10)
using_safely(.otherwise = NA_real_)$log("a")

using_quietly$sqrt(-1)
using_quietly$sqrt(4)

using_possibly$log("a", .otherwise = NA_real_)
using_possibly$log(10, .otherwise = NA_real_)
}
