% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_base.R
\name{vectorizing}
\alias{vectorizing}
\title{vectorizing tag to vectorize the input function}
\usage{
vectorizing(.vectorize.args = NULL, .SIMPLIFY = TRUE,
  .USE.NAMES = TRUE)
}
\arguments{
\item{.vectorize.args}{a character vector of arguments which should be
vectorized. Defaults to all arguments of FUN.}

\item{.SIMPLIFY}{logical or character string; attempt to reduce the result
to a vector, matrix or higher dimensional array; see the \code{simplify}
argument of \code{sapply}.}

\item{.USE.NAMES}{logical; use names if the first ... argument has names, or
if it is a character vector, use that character vector as the names.}
}
\description{
tag counterpart to \code{base::Vectorize}, and a wrapper around it.
}
\examples{
Vectorize(rep.int)(1:4, 4:1)
vectorizing$rep.int(1:4, 4:1)
}
