% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_side_effects.R
\name{viewing}
\alias{viewing}
\title{viewing tag to view the output of a call}
\usage{
viewing(.title = "View")
}
\arguments{
\item{.title}{itle for viewer window. Defaults to \code{"View"}}
}
\description{
viewing tag to view the output of a call
}
