% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_tracing_wrapping.R
\name{wrapping}
\alias{wrapping}
\title{wrapping tag for pre/post procession and composition}
\usage{
wrapping(.before = NULL, .after = NULL, .around = identity)
}
\arguments{
\item{.before}{an expression to run before the call}

\item{.after}{an expression to run after the call, can use \code{.} as the result
of the call}

\item{.around}{a function, or formula, to apply around the call before
evaluation, such as \code{capture.output} or \code{SuppressWarnings}}
}
\description{
wrapping tag for pre/post procession and composition
}
\examples{
wrapping(.before = print("hello"),
          .after = print(paste("result was:", .)),
          .around = suppressWarnings)$sqrt(-1)
}
