% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basics.R
\name{preserving_attr}
\alias{preserving_attr}
\title{preserving_attr tag to preserve attributes}
\usage{
preserving_attr(.arg = 1, incl_row.names = FALSE, incl_class = FALSE,
  incl_names = FALSE, incl_dim = FALSE)
}
\description{
preserving_attr tag to preserve attributes
}
\examples{
preserving_attr$map_dfr(head(iris,2),identity)
preserving_attr(incl_class = TRUE)$map_dfr(head(iris,2),identity)
}
