% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_tracing_wrapping.R
\name{tracing}
\alias{tracing}
\title{tracing tag}
\usage{
tracing(.tracer = NULL, .exit = NULL, .at = 1, .edit = FALSE)
}
\arguments{
\item{.tracer}{either a function or an unevaluated expression. The function
will be called or the expression will be evaluated either at the beginning
of the call, or before those steps in the call specified by the argument
at.}

\item{.exit}{either a function or an unevaluated expression. The function
will be called or the expression will be evaluated on exiting the function.}

\item{.at}{optional numeric vector or list. If supplied, \code{.tracer} will be
called just before the corresponding step in the body of the function.}

\item{.edit}{For complicated tracing, such as tracing within a loop inside
the function, you will need to insert the desired calls by editing the body
of the function. If so, supply the edit argument either as TRUE, or as the
name of the editor you want to use. Then \code{tracing()} will call edit and use the
version of the function after you edit it.}
}
\description{
Emulates the behavior of \code{base::trace}.
}
\details{
It's not a wrapper around \code{base::trace}, instead it modifies the body of the
input function. If the input function is an \emph{S3 generic} the effect of
the tracing can be overwritten by the argument values, better in this case
to use the relevant method as a function input.
}
\examples{
tracing(quote(print("hello")), .exit = quote(print("good bye")))$mean(1:5)
}
