% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_using_substituting.R
\name{using}
\alias{using}
\title{using tag to evaluate call in a given environment}
\usage{
using(.envir)
}
\arguments{
\item{.envir}{environment, list or vector on which to restrict the call}
}
\description{
This is the tag counterpart of \code{with} and it comes with a handy method for
\code{[}.
}
\examples{
using(iris)$ave(Sepal.Length,Species, FUN = mean)
vec <- c(1:2,a=2.5,3:4)
using(vec)[.>=a] # rather than ; vec[vec >= vec["a"]]
using(vec)[[.>=a]] # rather than ; vec >= vec["a"]
}
