% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{x}
\alias{x}
\title{fetch yaml
match top and bottom '---'}
\format{
An object of class \code{character} of length 5.
}
\usage{
x
}
\description{
HELPER functions (non-exported)
LEGACY (no longer used)
}
\details{
grep returns int vector indices
USAGE:   x  <-  c('---', 'a;lkdjf;asl ' , '999', 'af---', '---')
match_dashes(x)
}
\keyword{datasets}
